+++ apps/api/domain/pricing_calculator.go
@@ -13,7 +13,7 @@ func CalculatePrice(tiers []PricingTier, duration time.Duration) (PricingResult,
 	if len(tiers) == 0 {
 		return PricingResult{}, &Error{Type: BadRequest, Message: "snapshot has no tiers"}
 	}
-	durationMinutes := int(math.Ceil(duration.Minutes()))
+	durationMinutes := int64(math.Ceil(duration.Minutes()))
 	for _, tier := range tiers {
 		if tier.UpToMinutes >= durationMinutes {
 			return PricingResult{Price: tier.Price}, nil
