+++ apps/api/domain/variant_entity.go
@@ -5,7 +5,7 @@ import "time"
 type PricingTier struct {
 	Id          int64
 	VariantId   int64
-	UpToMinutes int
+	UpToMinutes int64
 	Price       float32
 	CreatedAt   time.Time
 }
